Implemented Phase 3. Updated compile.sh to include all th new files created. Updated monitor.c and treasure_hub.c so that the output from treasure_hunt.c goes through a pipe all the way to the hub, which then writes it to stdout.

+++ treasure_hub.c
@@ -6,8 +6,11 @@
 #include <signal.h>
 #include <errno.h>
 #include <fcntl.h>
+#include <limits.h>
 
+int monitor_pid=0;
 int monitor_status=0;//1 if running and 0 if not running
+int monitor_operation_finished=0;//1 if it finished the operation and 0 otherwise
 int accepting_commands=1;//1 if accepting and 0 if rejecting
 
 void print_commands(){
@@ -19,24 +22,27 @@ void print_commands(){
   printf("\t2 - list hunts\n");
   printf("\t3 - list treasures\n");
   printf("\t4 - view treasure\n");
-  printf("\t5 - stop monitor\n");
-  printf("\t6 - exit\n");
+  printf("\t5 - calculate scores\n");
+  printf("\t6 - stop monitor\n");
+  printf("\t7 - exit\n");
 }
 
 void monitor_stopped(int sig){
   /*
     handles SIGCHLD signal
   */
-  monitor_status=0;
   int ret_val,status;
   while ((ret_val=waitpid(-1, &status, WNOHANG))>0){//acknowledging terminated children
-    if(WIFEXITED(status)) printf("Monitor stopped with exit status %d\n",WEXITSTATUS(status));
+    if(ret_val==monitor_pid && WIFEXITED(status)){
+      monitor_status=0;
+      accepting_commands=1;
+      printf("Monitor stopped with exit status %d\n",WEXITSTATUS(status));
+    }
   }
   if(ret_val==-1 && errno!=ECHILD){
     perror(NULL);
     exit(-1);
   }
-  accepting_commands=1;
 }
 
 void monitor_started(int sig){
@@ -44,11 +50,15 @@ void monitor_started(int sig){
     handles SIGUSR1 signal (child sends SIGUSR1 when it started successfully)
   */
   monitor_status=1;
+  monitor_operation_finished=0;
   printf("Monitor started\n");
 }
 
-void do_nothing(){
-  //empty on purpose
+void monitor_finished_operation(){
+  /*
+    sets monitor_operation_finished to 1
+  */
+  monitor_operation_finished=1;
 }
 
 void remove_temporary_files(){
@@ -80,7 +90,7 @@ void setup_signals(){
     perror(NULL);
     exit(-1);
   }
-  sa.sa_handler = do_nothing;
+  sa.sa_handler = monitor_finished_operation;
   sa.sa_flags = SA_RESTART;
   if(sigaction(SIGUSR2, &sa, NULL) == -1){//child will send SIGUSR2 when it completed/acknowledged its task
     perror(NULL);
@@ -93,7 +103,7 @@ void setup_signals(){
   }
   sa.sa_handler = monitor_stopped;
   sa.sa_flags = SA_RESTART;
-  if(sigaction(SIGCHLD, &sa, NULL) == -1){//receives SIGCHLD when the monitor changes states
+  if(sigaction(SIGCHLD, &sa, NULL) == -1){//receives SIGCHLD when the a child changes states
     perror(NULL);
     exit(-1);
   }
@@ -120,22 +130,41 @@ void clear_buffer(){
   }
 }
 
-void start_monitor(int *monitor_pid,int *command_file_desc){
+void start_monitor(int *command_file_desc,int manager_pipes[2]){
   /*
     starts a new process for the treasure_monitor
   */
-  if(((*monitor_pid)=fork())<0){
+  
+  //initializes the pipes used for communication with treasure_hunt processes 
+  if(pipe(manager_pipes)==-1){
     perror(NULL);
     exit(-1);
   }
-  if(*monitor_pid==0){
+  
+  if((monitor_pid=fork())<0){
+    perror(NULL);
+    exit(-1);
+  }
+  if(monitor_pid==0){
     //child process
+    if(close(manager_pipes[0])==-1){//closes read end of the pipe
+      perror(NULL);
+      exit(-1);
+    }
+    if(dup2(manager_pipes[1],1)==-1){//links stdout to the write end of the pipe
+      perror(NULL);
+      exit(-1);
+    }
+    if(close(manager_pipes[1])==-1){//closes ORIGINAL DESCRIPTOR of the write end of the pipe
+      perror(NULL);
+      exit(-1);
+    }
     int temp_file_desc;
     if((temp_file_desc=open("./.monitor_command",O_CREAT|O_RDONLY,0644))==-1){
       perror(NULL);
       exit(-1);
     }
-    if(dup2(temp_file_desc,0)==-1){
+    if(dup2(temp_file_desc,0)==-1){//links stdin to the monitor_command file descriptor
       perror(NULL);
       exit(-1);
     }
@@ -144,6 +173,16 @@ void start_monitor(int *monitor_pid,int *command_file_desc){
       exit(-1);
     }
   }
+  //parent process
+  int flags;
+  if((flags=fcntl(manager_pipes[0],F_GETFL,0))==-1){//gets the flags of the read end of the pipe
+    perror(NULL);
+    exit(-1);
+  }
+  if(fcntl(manager_pipes[0],F_SETFL,flags|O_NONBLOCK)==-1){//sets reads from read end of the pipe to be non-blocking
+    perror(NULL);
+    exit(-1);
+  }
   if(((*command_file_desc)=open("./.monitor_command",O_CREAT|O_WRONLY|O_TRUNC,0644))==-1){
     perror(NULL);
     exit(-1);
@@ -151,7 +190,7 @@ void start_monitor(int *monitor_pid,int *command_file_desc){
 }
 
 int main(){
-  int monitor_pid,command_file_desc;
+  int command_file_desc,manager_pipes[2];
   setup_signals();
   while(1){
     print_commands();
@@ -175,24 +214,45 @@ int main(){
     }
     
     switch(command){
+      
       case 1:{
         if(!monitor_status){
 	  //start monitor
-	  start_monitor(&monitor_pid,&command_file_desc);
+	  start_monitor(&command_file_desc,manager_pipes);
 	  pause();
 	}
 	else printf("Monitor alreeady running (PID: %d)\n",monitor_pid);
         break;
       }
+	
       case 2:{
 	//list hunts
 	if(monitor_status){
 	  sigqueue(monitor_pid,SIGUSR1,(union sigval)2);
-	  pause();
+	  //reads from pipe
+	  char pipe_buffer[PIPE_BUF+1];
+	  while(1){
+	    int ret=read(manager_pipes[0],pipe_buffer,PIPE_BUF);
+	    if(ret==-1){
+	      if(errno!=EAGAIN && errno!=EINTR){//EAGAIN - no data available, EINTR - function inerrupted
+		perror(NULL);
+		exit(-1);
+	      }
+	      if(errno==EAGAIN && monitor_operation_finished==1){
+		monitor_operation_finished=0;
+		break;
+	      }
+	    }
+	    else{
+	      pipe_buffer[ret]='\0';
+	      printf("%s",pipe_buffer);
+	    }
+	  }
 	}
 	else printf("Monitor not running\n");
         break;
       }
+	
       case 3:{
 	//list treasures
         if(monitor_status){
@@ -222,11 +282,31 @@ int main(){
 	  }
 	  
 	  sigqueue(monitor_pid,SIGUSR1,(union sigval)3);
-	  pause();
+
+	  //reads from pipe
+	  char pipe_buffer[PIPE_BUF+1];
+	  while(1){
+	    int ret=read(manager_pipes[0],pipe_buffer,PIPE_BUF);
+	    if(ret==-1){
+	      if(errno!=EAGAIN && errno!=EINTR){//EAGAIN - no data available, EINTR - function inerrupted
+		perror(NULL);
+		exit(-1);
+	      }
+	      if(errno==EAGAIN && monitor_operation_finished==1){
+		monitor_operation_finished=0;
+		break;
+	      }
+	    }
+	    else{
+	      pipe_buffer[ret]='\0';
+	      printf("%s",pipe_buffer);
+	    }
+	  }
 	}
 	else printf("Monitor not running\n");
         break;
       }
+	
       case 4:{
 	// view treasure
         if(monitor_status){
@@ -314,12 +394,87 @@ int main(){
 	  }
 	  
 	  sigqueue(monitor_pid,SIGUSR1,(union sigval)4);
-	  pause();
+
+	  //reads from pipe
+	  char pipe_buffer[PIPE_BUF+1];
+	  while(1){
+	    int ret=read(manager_pipes[0],pipe_buffer,PIPE_BUF);
+	    if(ret==-1){
+	      if(errno!=EAGAIN && errno!=EINTR){//EAGAIN - no data available, EINTR - function inerrupted
+		perror(NULL);
+		exit(-1);
+	      }
+	      if(errno==EAGAIN && monitor_operation_finished==1){
+		monitor_operation_finished=0;
+		break;
+	      }
+	    }
+	    else{
+	      pipe_buffer[ret]='\0';
+	      printf("%s",pipe_buffer);
+	    }
+	  }
 	}
 	else printf("Monitor not running\n");
         break;
       }
+	
       case 5:{
+	//calculate scores
+	int score_pipes[2];
+	if(pipe(score_pipes)==-1){
+	  perror(NULL);
+	  exit(-1);
+	}
+	int child_pid;
+	if((child_pid=fork())==-1){
+	  perror(NULL);
+	  exit(-1);
+	}
+	if(child_pid==0){
+	  //child process
+	  if(close(score_pipes[0])==-1){//closes read end of the pipe
+	    perror(NULL);
+	    exit(-1);
+	  }
+	  if(dup2(score_pipes[1],1)==-1){//links write end of the pipe to stdout
+	    perror(NULL);
+	    exit(-1);
+	  }
+	  if(close(score_pipes[1])==-1){//closes ORIGINAL DESCRIPTOR of the write end of the pipe
+	    perror(NULL);
+	    exit(-1);
+	  }
+	  if(execl("./calculate_score_hub","./calculate_score_hub",NULL)==-1){
+	    perror(NULL);
+	    exit(-1);
+	  }
+	}
+	//parent process
+	if(close(score_pipes[1])==-1){//closes write end of the pipe
+	  perror(NULL);
+	  exit(-1);
+	}
+
+	//reads from pipe
+	char pipe_buffer[PIPE_BUF+1];
+	int ret;
+	while((ret=read(score_pipes[0],pipe_buffer,PIPE_BUF))>0){
+	  pipe_buffer[ret]='\0';
+	  printf("%s",pipe_buffer);
+	}
+	if(ret==-1){
+	  perror(NULL);
+	  exit(-1);
+	}
+	if(close(score_pipes[0])==-1){
+	  perror(NULL);
+	  exit(-1);
+	}
+	break;
+      }
+	
+      case 6:{
 	//stop monitor
         if(monitor_status){
 	  if(kill(monitor_pid,SIGUSR2)!=0){
@@ -335,6 +490,10 @@ int main(){
 	    perror(NULL);
 	    exit(-1);
 	  }
+	  if(close(manager_pipes[0])==-1 || close(manager_pipes[1])==-1){
+	    perror(NULL);
+	    exit(-1);
+	  }
 	  printf("Stopping monitor\n");
 	  accepting_commands=0;
 	}
@@ -343,16 +502,19 @@ int main(){
 	}
         break;
       }
-      case 6:{
+	
+      case 7:{
 	//exit
 	if(monitor_status) printf("Monitor still running (PID: %d)\n",monitor_pid);
 	else return 0;
 	break;
       }
+	
       default:{
         printf("Invalid command\n");
         break;
       }
+	
     }
   }
   return 0;
